+++ app/src/main/java/in/basulabs/shakealarmclock/backend/AlarmBroadcastReceiver.java
@@ -14,7 +14,7 @@ GNU Affero General Public License for more details.
 You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-package in.basulabs.shakealarmclock.frontend;
+package in.basulabs.shakealarmclock.backend;
 
 import static android.content.Context.POWER_SERVICE;
