+++ app/src/main/java/in/basulabs/shakealarmclock/backend/Service_RingAlarm.java
@@ -65,7 +65,6 @@ import in.basulabs.audiofocuscontroller.AudioFocusController;
 import in.basulabs.shakealarmclock.R;
 import in.basulabs.shakealarmclock.frontend.Activity_AlarmsList;
 import in.basulabs.shakealarmclock.frontend.Activity_RingAlarm;
-import in.basulabs.shakealarmclock.frontend.AlarmBroadcastReceiver;
 
 public class Service_RingAlarm extends Service implements SensorEventListener,
 	AudioFocusController.OnAudioFocusChangeListener {
