+++ app/src/main/java/in/basulabs/shakealarmclock/backend/Service_SetAlarmsPostBoot.java
@@ -49,7 +49,6 @@ import java.util.concurrent.atomic.AtomicReference;
 
 import in.basulabs.shakealarmclock.R;
 import in.basulabs.shakealarmclock.frontend.Activity_RequestPermIntro;
-import in.basulabs.shakealarmclock.frontend.AlarmBroadcastReceiver;
 
 public class Service_SetAlarmsPostBoot extends Service {
