+++ app/src/main/java/in/basulabs/shakealarmclock/backend/Worker_ActivateAlarms.java
@@ -41,7 +41,6 @@ import java.util.concurrent.atomic.AtomicReference;
 
 import in.basulabs.shakealarmclock.R;
 import in.basulabs.shakealarmclock.frontend.Activity_RequestPermIntro;
-import in.basulabs.shakealarmclock.frontend.AlarmBroadcastReceiver;
 
 public class Worker_ActivateAlarms extends Worker {
