+++ app/src/main/java/in/basulabs/shakealarmclock/frontend/Activity_IntentManager.java
@@ -64,6 +64,7 @@ import java.util.Collections;
 import java.util.Objects;
 import java.util.concurrent.atomic.AtomicInteger;
 
+import in.basulabs.shakealarmclock.backend.AlarmBroadcastReceiver;
 import in.basulabs.shakealarmclock.backend.AlarmDatabase;
 import in.basulabs.shakealarmclock.backend.AlarmEntity;
 import in.basulabs.shakealarmclock.backend.ConstantsAndStatics;
